+++ web-admin/src/components/deployments/ProjectBuilding.svelte
