+++ web-common/src/components/icons/CancelCircleInverse.svelte
